+++ spec/services/doc_auth/passports/dos/responses/health_check_failure_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe DocAuth::Passports::HealthCheckResponseFailure do
+RSpec.describe DocAuth::Passports::Dos::Responses::HealthCheckFailure do
   subject(:health_check_result) { described_class.new(faraday_error) }
 
   let(:health_check_endpoint) do
